+++ util/matrix.go
@@ -1,11 +1,11 @@
 package util
 
-func CreateMatrix(size int) [][]float64 {
-	matrix := make([][]float64, size)
+func CreateMatrix(size int) [][]float32 {
+	matrix := make([][]float32, size)
 	for i := range size {
-		matrix[i] = make([]float64, size)
+		matrix[i] = make([]float32, size)
 		for j := range size {
-			matrix[i][j] = float64(i + j)
+			matrix[i][j] = float32(i + j)
 		}
 	}
 	return matrix
